% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabocha.R
\name{cabochaFlatXML}
\alias{cabochaFlatXML}
\title{Call cabocha -f3 command}
\usage{
cabochaFlatXML(texts, rcpath = NULL, as.tibble = FALSE,
  force.utf8 = FALSE)
}
\arguments{
\item{texts}{characters that you want to pass to CaboCha.}

\item{rcpath}{fullpath to MECABRC if any.}

\item{as.tibble}{boolean. if false, then return flatXML dataframe.}

\item{force.utf8}{boolean. If true, it will read cabocha output xml with UTF-8.}
}
\value{
flat XML made by \code{flatxml::fxml_importXMLFlat(CaboChaOutputXML)}
}
\description{
Call cabocha -f3 command via shell(), then return results as a flat XML.
It requires that CaboCha has already been installed and available.
}
\seealso{
\url{https://www.rdocumentation.org/packages/flatxml/versions/0.0.2}
}
