% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CabochaR}
\alias{CabochaR}
\title{Utility for parsing CaboCha output}
\usage{
CabochaR(fxml)
}
\arguments{
\item{fxml}{flat XML that comes out from \code{cabochaFlatXML(as.tibble = FALSE)}}
}
\value{
R6 class object having fields and methods below.
\itemize{
  \item attributes: tibble. Each rows consist of the element id and attributes of chunk.
  \item morphs: list of tibbles. Each rows represent morphemes of which the chunk is made.
  \item as_tibble(): return a tibble compatible with CabochaR output.
}
}
\description{
convert flat XML into CabochaR compatible output.
}
\seealso{
\url{https://minowalab.org/cabochar/}
}
