% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-rmecab.R
\name{gbs_as_tokens}
\alias{gbs_as_tokens}
\title{Pack prettified output as quanteda tokens}
\usage{
gbs_as_tokens(df, pull = "token", n = 1L, sep = "-", what = "fastestword", ...)
}
\arguments{
\item{df}{A prettified data.frame of tokens.}

\item{pull}{Column to be packed into text or ngrams body. Default value is `token`.}

\item{n}{Integer passed to \code{quanteda::char_ngrams()}.}

\item{sep}{Character scalar internally used as the concatenator of ngrams
inside \code{quanteda::char_ngrams}.}

\item{what}{Character scalar; which tokenizer to use in \code{quanteda::tokens()}.}

\item{...}{Other arguments are passed to \code{quanteda::tokens()}.}
}
\value{
A quanteda 'token' class object.
}
\description{
Pack prettified output as quanteda tokens
}
\examples{
xml <- ppn_parse_xml(system.file("sample.xml", package = "pipian"))
gbs_as_tokens(xml)
}
\seealso{
Other gibasa: 
\code{\link{gbs_collocate}()},
\code{\link{gbs_c}()},
\code{\link{gbs_dfm}()},
\code{\link{gbs_freq}()}
}
\concept{gibasa}
