% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-rmecab.R
\name{gbs_freq}
\alias{gbs_freq}
\title{An alternative of RMeCabFreq}
\usage{
gbs_freq(df, ..., .name_repair = TRUE)
}
\arguments{
\item{df}{Prettified data.frame of tokenized sentences.}

\item{...}{Other arguments are passed to \code{dplyr::tally()}.}

\item{.name_repair}{Logical.
If true, then rename column names to RMecabFreq compatible style.}
}
\value{
tibble
}
\description{
An alternative of RMeCabFreq
}
\seealso{
Other gibasa: 
\code{\link{gbs_as_tokens}()},
\code{\link{gbs_collocate}()},
\code{\link{gbs_c}()},
\code{\link{gbs_dfm}()}
}
\concept{gibasa}
