% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_cabocha_available}
\alias{is_cabocha_available}
\title{Check if cabocha command is available}
\usage{
is_cabocha_available()
}
\value{
Logical.
}
\description{
Check if cabocha command is available
}
\keyword{internal}
