% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_xml}
\alias{parse_xml}
\title{Parse XML output of CaboCha}
\usage{
parse_xml(path)
}
\arguments{
\item{path}{String scalar.}
}
\value{
data.frame.
}
\description{
Parse XML output of CaboCha
}
\keyword{internal}
