% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pluckAttrs}
\alias{pluckAttrs}
\title{Utility for parsing CaboCha output}
\usage{
pluckAttrs(str, idx)
}
\arguments{
\item{str}{strings representing feature attribute of CaboCha output tok.}

\item{idx}{index to pluck.}
}
\description{
Pluck a value from feature attribute of CaboCha output tok.
}
