% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabocha-parser.R
\name{ppn_as_tokenindex}
\alias{ppn_as_tokenindex}
\title{Cast xml as data.table}
\usage{
ppn_as_tokenindex(path)
}
\arguments{
\item{path}{Character vector came from \code{ppn_cabocha}.}
}
\value{
data.table retured from \code{rsyntax::as_tokenindex}.
}
\description{
Cast xml as data.table
}
