% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabocha-exec.R
\name{ppn_cabocha}
\alias{ppn_cabocha}
\title{Execute cabocha command}
\usage{
ppn_cabocha(text, rcpath = NULL)
}
\arguments{
\item{text}{Characters vector to be analyzed by CaboCha.}

\item{rcpath}{String scalar that path to the `mecabrc` file if any.}
}
\value{
Character vector that path to result xml files.
}
\description{
Execute cabocha -f3 -n1 command using \code{processx::run},
then return path to the result files.
}
