% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabocha-parser.R
\name{ppn_make_graph}
\alias{ppn_make_graph}
\title{Cast dependency structure as an igraph}
\usage{
ppn_make_graph(df)
}
\arguments{
\item{df}{Output of \code{pipian::ppn_parse_xml}.}
}
\value{
An 'igraph' object is returned.
}
\description{
Cast dependency structure as an igraph
}
\examples{
xml <- ppn_parse_xml(system.file("sample.xml", package = "pipian"))
ppn_make_graph(xml)
}
