% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabocha-parser.R
\name{ppn_parse_xml}
\alias{ppn_parse_xml}
\title{Parse xml output of CaboCha}
\usage{
ppn_parse_xml(
  path,
  into = c("POS1", "POS2", "POS3", "POS4", "X5StageUse1", "X5StageUse2", "Original",
    "Yomi1", "Yomi2")
)
}
\arguments{
\item{path}{Character vector came from \code{ppn_cabocha}.}

\item{into}{Character vector. Feature names of output.}
}
\value{
data.frame
}
\description{
Parse xml output of CaboCha
}
