% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabocha-parser.R
\name{ppn_plot_igraph}
\alias{ppn_plot_igraph}
\title{Plot dependency strutcure using igraph}
\usage{
ppn_plot_igraph(df, sentence_id)
}
\arguments{
\item{df}{Output of \code{ppn_parse_xml}.}

\item{sentence_id}{Sentence id to be kept in igraph.}
}
\description{
Plot dependency strutcure using igraph
}
