% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabocha.R
\name{CabochaTbl}
\alias{CabochaTbl}
\title{Call cabocha -f3 command}
\usage{
CabochaTbl(texts, rcpath = NULL, force.utf8 = FALSE)
}
\arguments{
\item{texts}{characters that you want to pass to CaboCha.}

\item{rcpath}{fullpath to MECABRC if any.}

\item{force.utf8}{boolean. If true, it will read cabocha output xml with UTF-8.}
}
\value{
R6 class object having fields and methods below.
\itemize{
  \item texts
  \item rcpath
  \item tbl: tibble that each rows represents a phrase.
  \item tbl2graph(): convert tbl to igraph graph object.
  \item plot(): simplely plot the syntactic tree.
}
}
\description{
Call cabocha -f3 command via system(), then return results as a tibble.
It requires that CaboCha has already been installed and available.
}
