% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_cabocha_available}
\alias{is_cabocha_available}
\title{Check if cabocha command is available}
\usage{
is_cabocha_available(version = "0.69")
}
\arguments{
\item{version}{version of CaboCha available.}
}
\value{
logical
}
\description{
Check if cabocha command is available
}
