% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipian-package.R
\docType{package}
\name{pipian}
\alias{pipian}
\alias{pipian-package}
\title{pipian: Tiny Interface to CaboCha for R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A tiny parser of CaboCha output XML powered by 'Rcpp'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/pipian}
  \item Report bugs at \url{https://github.com/paithiov909/pipian/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Authors:
\itemize{
  \item Marcin Kalicinski (Author of rapidxml)
}

}
\keyword{internal}
