% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabocha-parser.R
\name{ppn_as_tokenindex}
\alias{ppn_as_tokenindex}
\title{Cast xml as data.table}
\usage{
ppn_as_tokenindex(tokens)
}
\arguments{
\item{tokens}{Output from \code{pipian::ppn_parse_xml}.}
}
\value{
data.table returned from \code{rsyntax::as_tokenindex}.
}
\description{
Cast xml as data.table
}
\examples{
xml <- ppn_parse_xml(system.file("sample.xml", package = "pipian"))
ppn_as_tokenindex(xml)
}
