% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabocha-parser.R
\name{ppn_as_tokenindex}
\alias{ppn_as_tokenindex}
\title{Cast xml as data.table}
\usage{
ppn_as_tokenindex(tokens)
}
\arguments{
\item{tokens}{Output from \code{pipian::ppn_parse_xml}.}
}
\value{
data.table retured from \code{rsyntax::as_tokenindex}.
}
\description{
Cast xml as data.table
}
