% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabocha-parser.R
\name{ppn_as_tokenindex}
\alias{ppn_as_tokenindex}
\title{Cast XML as data.table}
\usage{
ppn_as_tokenindex(tokens)
}
\arguments{
\item{tokens}{Output from \code{pipian::ppn_parse_xml}.}
}
\value{
A data.table casted with \code{rsyntax::as_tokenindex} is returned.
}
\description{
Cast XML as data.table
}
\examples{
xml <- ppn_parse_xml(system.file("sample.xml", package = "pipian"))
head(ppn_as_tokenindex(xml))
}
