% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabocha-exec.R
\name{ppn_cabocha}
\alias{ppn_cabocha}
\title{Execute cabocha command}
\usage{
ppn_cabocha(text, rcpath = NULL)
}
\arguments{
\item{text}{A character vector to be parsed with CaboCha.}

\item{rcpath}{String; path to the `mecabrc` file if any.}
}
\value{
Paths to the CaboCha XML output are returned.
}
\description{
Execute cabocha -f3 -n1 command using \code{processx::run},
then return the paths to the temporary XML files.
}
\examples{
\dontrun{
  ppn_cabocha(enc2utf8("\u96e8\u306b\u3082\u8ca0\u3051\u305a"))
}
}
