% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabocha-parser.R
\name{ppn_parse_xml}
\alias{ppn_parse_xml}
\title{Parse XML output of CaboCha}
\usage{
ppn_parse_xml(
  path,
  into = c("POS1", "POS2", "POS3", "POS4", "X5StageUse1", "X5StageUse2", "Original",
    "Yomi1", "Yomi2"),
  col_select = seq_along(into)
)
}
\arguments{
\item{path}{String; output from \code{pipian::ppn_cabocha}.}

\item{into}{Character vector; feature names of output.}

\item{col_select}{Character or integer vector; features that will be kept
in the result.}
}
\value{
A data.frame.
}
\description{
Parse XML output of CaboCha
}
\examples{
head(ppn_parse_xml(system.file("sample.xml", package = "pipian")))
}
