% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabocha-parser.R
\name{ppn_parse_xml}
\alias{ppn_parse_xml}
\title{Parse xml output of CaboCha}
\usage{
ppn_parse_xml(
  path,
  into = c("POS1", "POS2", "POS3", "POS4", "X5StageUse1", "X5StageUse2", "Original",
    "Yomi1", "Yomi2")
)
}
\arguments{
\item{path}{String; output from \code{pipian::ppn_cabocha}.}

\item{into}{Character vector; the feature names of output.}
}
\value{
A data.frame.
}
\description{
Parse xml output of CaboCha
}
\examples{
head(ppn_parse_xml(system.file("sample.xml", package = "pipian")))
}
