% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabocha-parser.R
\name{ppn_plot_igraph}
\alias{ppn_plot_igraph}
\title{Plot dependency structure using igraph}
\usage{
ppn_plot_igraph(df)
}
\arguments{
\item{df}{Output of \code{pipian::ppn_parse_xml}.}
}
\value{
The graph object is returned invisibly.
}
\description{
Plot dependency structure using igraph
}
\examples{
xml <- ppn_parse_xml(system.file("sample.xml", package = "pipian"))
if (FALSE) {
  ppn_plot_igraph(xml)
}
}
