% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reset_encoding}
\alias{reset_encoding}
\title{Format Character Vector}
\usage{
reset_encoding(vec, enc = "UTF-8")
}
\arguments{
\item{vec}{Character vector.}

\item{enc}{Encoding of vec.}
}
\description{
Format Character Vector
}
\keyword{internal}
